requirements.txt
# Environment
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
openenv-core==0.2.3

# HTTP
requests>=2.25.0
httpx>=0.24.0

# LLM + Agents
openai>=1.0.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
